Scene de Scoring
La scene de scoring affiche le nombre de preuve trouvée et le nom d'essais restant ainsi que des étoiles selon le nombre de fautes faites
Ramène à l'écran de selection de Level ajustant le score du Level

+++ Assets/Script/Level.cs
@@ -9,7 +9,7 @@ public class Level : MonoBehaviour
     public string myDescription;
     public string myScene;
     public int myNumber;
-    public int myScore;
+    public float myScore;
 
     public RectTransform myTransform;
